+++ actions/null_actions.js
@@ -1,0 +1,15 @@
+var types = require('./types.js');
+
+module.exports = {};
+
+function process_tip(msg){
+    var responses = new types.Responses();
+    return responses;
+}
+module.exports.process_tip = process_tip;
+
+function process_chat(msg) {
+    var responses = new types.Responses();
+    return responses;
+}
+module.exports.process_chat = process_chat;
