+++ package.json
@@ -25,10 +25,13 @@
     "MFCAuto": "zombiealex/MFCAuto",
     "config": "^1.21.0",
     "config-heroku": "^0.1.1-a",
+    "express": "^4.14.0",
     "foreman": "^1.4.1",
     "mathjs": "^3.4.1",
     "minimist": "^1.2.0",
-    "moment": "^2.14.1"
+    "moment": "^2.14.1",
+    "morgan": "^1.7.0",
+    "serialport": "^4.0.1"
   },
   "engines": {
     "node": "4.4.7"
